+++ app/javascript/packs/hello_vue.js
@@ -2,27 +2,3 @@ import Vue from 'vue/dist/vue.esm'
 import 'bootstrap/dist/css/bootstrap.css'
 import 'bootstrap-vue/dist/bootstrap-vue.css'
 Vue.config.productionTip = false
-
-var app = new Vue ({
-  el: '#app',
-  data: {
-    newItem: '',
-    todos: []
-  },
-  methods: {
-    addItem: function(event) {
-      if(this.newItem == '') return;
-      var todo = {
-        item: this.newItem,
-        isDone: false
-      };
-
-      this.todos.push(todo)
-      this.newItem=''
-    },
-    deleteItem: function(index) {
-      // alert(index);
-      this.todos.splice(index, 1)
-    }
-  }
-})
