+++ app/views/page/home.html.erb
@@ -1,28 +1,11 @@
-<div id="app">
-  <h2>TODO List</h2>
-  <form v-on:submit.prevent>
-    <input class="form-control mb-2 mr-sm-2" type="text" v-model="newItem">
-    <button class="btn-primary" v-on:click="addItem">
-      Add
-    </button>
-  </form>
-  <ul>
-    <li v-for="(todo, index) in todos">
-      <input class="form-check-input" type="checkbox" v-model="todo.isDone">
-      <span v-bind:class="{ done: todo.isDone }">
-        {{ todo.item }}
-      </span>
-      <button  class="btn-secondary" v-on:click="deleteItem(index)">
-        Delete
-      </button>
-    </li>
-  </ul>
-  <pre>{{ $data }}</pre>
-
-  <%= link_to 'books#index', books_path %>
-  <%= link_to 'page#basic', basic_page_index_path %>
-  <%= link_to 'page#bitcoin', bitcoin_page_index_path %>
+<h2>Vue.js Practice Links</h2>
+<hr>
+<div class="d-flex flex-row">
+  <div class="p-2"><%= link_to 'books#index', books_path %></div>
+  <div class="p-2"><%= link_to 'page#todo', todo_page_index_path %></div>
+  <div class="p-2"><%= link_to 'page#basic', basic_page_index_path %></div>
+  <div class="p-2"><%= link_to 'page#bitcoin', bitcoin_page_index_path %></div>
+  <div class="p-2"><%= link_to 'page#template', template_page_index_path %></div>
 </div>
-
 
 <%= javascript_pack_tag 'hello_vue' %>
