+++ app/views/page/template.html.erb
@@ -1,0 +1,16 @@
+<div id="template">
+
+  <h2>テンプレート構文 プラクティス</h2>
+  <hr>
+  <p>
+    {{ message}}
+  </p>
+  <button v-on:click='clickhandler'>
+    change!
+  </button>
+
+</div>
+<br><br><br>
+<%= link_to '>>HOME', root_path %>
+
+<%= javascript_pack_tag 'hello_vue_template' %>
