+++ config/routes.rb
@@ -3,8 +3,10 @@ Rails.application.routes.draw do
   resources :page, only: :index do
     collection do
       get 'home'
+      get 'todo'
       get 'basic'
       get 'bitcoin'
+      get 'template'
     end
   end
   resources :books, only: :index
